+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+ - `Currency.all` raises a more helpful error message
+   (`Currency::MissingAttributeError`)if a currency has no priority
  - `Currency` implements `Enumerable`.
  - `Currency#<=>` sorts alphabetically by `id` if the `priority`s are the same,
    and no longer raises an error if one of the priorities is missing.